Update README.md
Built on the colossal glory of the previous edition of SAUDI DRONES SUMMIT AND EXPO 2019, we are honored to bring the “2nd Annual Saudi Drones Summit and Expo”, in the capital city of Kingdom- Riyadh, to answer most pressing challenges with a broader scope and provide a glimpse at what’s coming next in the race. Keeping with the forefront of industry innovations and market requirements the format of the event will range from multi-track conference program to hands-on workshops, keynote sessions, panel discussions, networking opportunities and more exhibitors than before making it the best choice for professionals who need to learn how to best use UAV to maximize their organization’s efficiency, save money, improve safety and mitigate risk. Supporting the Kingdom’s VISION 2030 and General authority for civil aviation, Saudi Arabia commendable initiatives for licensing drones to boost technology competence of kingdom, this high-profile event is a multi-stakeholder event ideally positioned to feature premium assortment of 300+ attendees, 30+ Industry Experts, 40+ sponsors and exhibitors and 14+ hours of networking